antimalaria_backend_model/requirements.txt
absl-py==2.3.0
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.4
billiard==4.2.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.3
Django==5.2.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
filelock==3.18.0
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.72.1
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
joblib==1.5.1
keras==3.10.0
kombu==5.5.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psycopg2-binary==2.9.10
PubChemPy==1.0.4
pycparser==2.22
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
rdkit==2025.3.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.2
lightgbm